Split digitalocean module into submodules
